+++ src/components/notification/Notification.vue
@@ -9,6 +9,7 @@
                 class="delete"
                 type="button"
                 @click="close"
+                :aria-label="ariaCloseLabel"
             />
             <div class="media">
                 <div v-if="icon && hasIcon" class="media-left">
@@ -16,7 +17,8 @@
                         :icon="icon"
                         :pack="iconPack"
                         both
-                        size="is-large"/>
+                        size="is-large"
+                        aria-hidden/>
                 </div>
                 <div class="media-content">
                     <slot/>
@@ -31,6 +33,9 @@
 
     export default {
         name: 'BNotification',
-        mixins: [MessageMixin]
+        mixins: [MessageMixin],
+        props: {
+            ariaCloseLabel: ''
+        }
     }
 </script>