+++ src/components/table/Table.vue
@@ -69,7 +69,12 @@
                             }]"
                             @click="selectRow(row)"
                             @dblclick="$emit('dblclick', row)"
-                            @contextmenu="$emit('contextmenu', row, $event)">
+                            @contextmenu="$emit('contextmenu', row, $event)"
+                            :draggable="draggable"
+                            @dragstart="handleDragStart($event, row, index)"
+                            @drop="handleDrop($event, row, index)"
+                            @dragover="handleDragOver($event, row, index)"
+                            @dragleave="handleDragLeave($event, row, index)">
 
                             <td
                                 v-if="showDetailRowIcon"
@@ -267,7 +272,12 @@
             },
             iconPack: String,
             mobileSortPlaceholder: String,
-            customRowKey: String
+            customRowKey: String,
+            draggable: {
+                type: Boolean,
+                defualt: false
+            }
+
         },
         data() {
             return {
@@ -731,6 +741,30 @@
                         this.sort(column, true)
                     }
                 })
+            },
+            /**
+             * Emits drag start event
+             */
+            handleDragStart(event, row, index) {
+                this.$emit('dragstart', {event, row, index})
+            },
+            /**
+             * Emits drop event
+             */
+            handleDrop(event, row, index) {
+                this.$emit('drop', {event, row, index})
+            },
+            /**
+             * Emits drag over event
+             */
+            handleDragOver(event, row, index) {
+                this.$emit('dragover', {event, row, index})
+            },
+            /**
+             * Emits drag leave event
+             */
+            handleDragLeave(event, row, index) {
+                this.$emit('dragleave', {event, row, index})
             }
         },