+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### New features
 
+* Add ``custom-formatter`` prop to autocomplete
+* Add ``draggable`` prop to table (thanks @suneab)
+* Improve accessibility for notification (thanks @mateuswetah)
+
 ### Fixes
 
 * #1251 Fix Datepicker closing when page back to ``min-date`` month